split elf and macho linking into their own files
